requirements.txt
rembg>=2.0.59
Pillow>=10.4.0
onnxruntime>=1.23.0
